Fix after refactor.
